Refactor code structure
Refactored the codebase for improved organization and maintainability.
[skip gpt_engineer]
