+++ src/components/CodePlayground.tsx
@@ -1,39 +1,18 @@
 import React, { useState, useEffect, useRef } from 'react';
 import Editor from '@monaco-editor/react';
 import { Card, CardContent, CardHeader, CardTitle } from './ui/card';
-import { Button } from './ui/button';
-import { Copy, Download, Play, Maximize2 } from 'lucide-react';
 import { useToast } from '@/hooks/use-toast';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from './ui/select';
 import { ResizablePanel, ResizablePanelGroup } from './ui/resizable';
 import { executeJavaScript, executeHTML } from '@/utils/codeExecutor';
+import { PlaygroundToolbar } from './playground/PlaygroundToolbar';
+import { PlaygroundOutput } from './playground/PlaygroundOutput';
+import { usePopoutWindow } from '@/hooks/usePopoutWindow';
+import { SUPPORTED_LANGUAGES } from './playground/constants';
 
 interface CodePlaygroundProps {
   defaultLanguage?: string;
   defaultValue?: string;
 }
-
-const SUPPORTED_LANGUAGES = [
-  { value: 'javascript', label: 'JavaScript', extension: '.js' },
-  { value: 'typescript', label: 'TypeScript', extension: '.ts' },
-  { value: 'python', label: 'Python', extension: '.py' },
-  { value: 'java', label: 'Java', extension: '.java' },
-  { value: 'csharp', label: 'C#', extension: '.cs' },
-  { value: 'cpp', label: 'C++', extension: '.cpp' },
-  { value: 'go', label: 'Go', extension: '.go' },
-  { value: 'rust', label: 'Rust', extension: '.rs' },
-  { value: 'ruby', label: 'Ruby', extension: '.rb' },
-  { value: 'php', label: 'PHP', extension: '.php' },
-  { value: 'sql', label: 'SQL', extension: '.sql' },
-  { value: 'html', label: 'HTML', extension: '.html' },
-  { value: 'css', label: 'CSS', extension: '.css' },
-  { value: 'json', label: 'JSON', extension: '.json' },
-  { value: 'markdown', label: 'Markdown', extension: '.md' },
-  { value: 'hcl', label: 'Terraform', extension: '.tf' },
-  { value: 'bicep', label: 'Bicep', extension: '.bicep' },
-  { value: 'powershell', label: 'PowerShell', extension: '.ps1' },
-  { value: 'shell', label: 'Bash/Shell', extension: '.sh' },
-];
 
 const CodePlayground: React.FC<CodePlaygroundProps> = ({
   defaultLanguage = 'javascript',
@@ -42,114 +21,26 @@ const CodePlayground: React.FC<CodePlaygroundProps> = ({
   const [code, setCode] = useState(defaultValue);
   const [language, setLanguage] = useState(defaultLanguage);
   const [output, setOutput] = useState<string>('');
-  const [isOutputPopped, setIsOutputPopped] = useState(false);
   const { toast } = useToast();
   const outputRef = useRef<HTMLDivElement>(null);
   const iframeRef = useRef<HTMLDivElement>(null);
-  const popoutWindowRef = useRef<Window | null>(null);
+  
+  const { isOutputPopped, handlePopOutput } = usePopoutWindow(code, language, output);
 
   useEffect(() => {
-    if (popoutWindowRef.current && !popoutWindowRef.current.closed) {
-      const doc = popoutWindowRef.current.document;
-      const outputElement = doc.getElementById('output');
-      const isDark = document.documentElement.classList.contains('dark');
-      
-      if (outputElement) {
-        if (language === 'html') {
-          const iframe = executeHTML(code);
-          iframe.classList.toggle('dark-iframe', isDark);
-          outputElement.innerHTML = '';
-          outputElement.appendChild(iframe);
-        } else {
-          outputElement.innerHTML = `<pre class="whitespace-pre-wrap font-mono p-4 ${isDark ? 'text-white' : 'text-black'}">${output}</pre>`;
-        }
-      }
+    const savedCode = localStorage.getItem('playground-code');
+    const savedLanguage = localStorage.getItem('playground-language');
+    
+    if (savedCode) {
+      setCode(savedCode);
+      localStorage.removeItem('playground-code');
     }
-  }, [output, language, code]);
-
-  const handlePopOutput = () => {
-    if (isOutputPopped && popoutWindowRef.current && !popoutWindowRef.current.closed) {
-      popoutWindowRef.current.close();
-      setIsOutputPopped(false);
-      return;
+    
+    if (savedLanguage && SUPPORTED_LANGUAGES.some(lang => lang.value === savedLanguage)) {
+      setLanguage(savedLanguage);
+      localStorage.removeItem('playground-language');
     }
-
-    const isDark = document.documentElement.classList.contains('dark');
-    const popoutWindow = window.open('', 'CodeOutput', 'width=600,height=400,resizable=yes');
-    if (popoutWindow) {
-      popoutWindow.document.write(`
-        <!DOCTYPE html>
-        <html class="${isDark ? 'dark' : ''}">
-          <head>
-            <title>Code Output</title>
-            <style>
-              body { 
-                margin: 0;
-                padding: 0;
-                font-family: monospace;
-              }
-              body.dark {
-                background: #1e1e1e;
-                color: #fff;
-              }
-              body:not(.dark) {
-                background: #ffffff;
-                color: #000;
-              }
-              #output {
-                height: 100vh;
-                width: 100%;
-              }
-              #output iframe {
-                width: 100%;
-                height: 100%;
-                border: none;
-              }
-              #output iframe.dark-iframe {
-                background: #1e1e1e;
-                color: #fff;
-              }
-              #output iframe:not(.dark-iframe) {
-                background: #ffffff;
-                color: #000;
-              }
-              pre {
-                margin: 0;
-                padding: 1rem;
-                white-space: pre-wrap;
-                word-wrap: break-word;
-              }
-            </style>
-          </head>
-          <body class="${isDark ? 'dark' : ''}">
-            <div id="output">
-              ${language === 'html' 
-                ? ''  // Will be populated by the useEffect
-                : `<pre class="whitespace-pre-wrap font-mono p-4 ${isDark ? 'text-white' : 'text-black'}">${output}</pre>`}
-            </div>
-          </body>
-        </html>
-      `);
-      popoutWindow.document.close();
-      popoutWindowRef.current = popoutWindow;
-      setIsOutputPopped(true);
-
-      if (language === 'html') {
-        const iframe = executeHTML(code);
-        const outputElement = popoutWindow.document.getElementById('output');
-        iframe.classList.toggle('dark-iframe', isDark);
-        if (outputElement) {
-          outputElement.innerHTML = '';
-          outputElement.appendChild(iframe);
-        }
-      }
-
-      popoutWindow.onbeforeunload = () => {
-        setIsOutputPopped(false);
-        return null;
-      };
-    }
-  };
+  }, []);
 
   const handleCopy = async () => {
     try {
@@ -230,60 +121,16 @@ const CodePlayground: React.FC<CodePlaygroundProps> = ({
   return (
     <Card className="w-full h-[90vh] mx-auto bg-card shadow-lg">
       <CardHeader className="border-b border-border/20">
-        <CardTitle className="flex justify-between items-center">
-          <div className="flex items-center gap-4">
-            <span className="text-xl font-semibold text-card-foreground">Code Playground</span>
-            <Select value={language} onValueChange={setLanguage}>
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="Select language" />
-              </SelectTrigger>
-              <SelectContent>
-                {SUPPORTED_LANGUAGES.map((lang) => (
-                  <SelectItem key={lang.value} value={lang.value}>
-                    {lang.label}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-          <div className="flex gap-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleRun}
-              className="hover:bg-accent"
-            >
-              <Play className="w-4 h-4 mr-2" />
-              Run
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handlePopOutput}
-              className="hover:bg-accent"
-            >
-              <Maximize2 className="w-4 h-4 mr-2" />
-              {isOutputPopped ? 'Close Output' : 'Pop Output'}
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleSaveToFile}
-              className="hover:bg-accent"
-            >
-              <Download className="w-4 h-4 mr-2" />
-              Save
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleCopy}
-              className="hover:bg-accent"
-            >
-              <Copy className="w-4 h-4 mr-2" />
-              Copy
-            </Button>
-          </div>
+        <CardTitle>
+          <PlaygroundToolbar
+            language={language}
+            onLanguageChange={setLanguage}
+            onRun={handleRun}
+            onPopOutput={handlePopOutput}
+            onSaveToFile={handleSaveToFile}
+            onCopy={handleCopy}
+            isOutputPopped={isOutputPopped}
+          />
         </CardTitle>
       </CardHeader>
       <CardContent className="p-4 pb-8 h-[calc(90vh-5rem)]">
@@ -311,15 +158,12 @@ const CodePlayground: React.FC<CodePlaygroundProps> = ({
           </ResizablePanel>
           {!isOutputPopped && (
             <ResizablePanel defaultSize={40}>
-              <div className="h-full flex flex-col">
-                {language === 'html' ? (
-                  <div ref={iframeRef} className={`w-full h-full ${document.documentElement.classList.contains('dark') ? 'bg-[#1e1e1e]' : 'bg-white'}`} />
-                ) : (
-                  <div ref={outputRef} className="w-full h-full p-4 font-mono text-sm overflow-auto bg-black text-white">
-                    {output}
-                  </div>
-                )}
-              </div>
+              <PlaygroundOutput
+                language={language}
+                output={output}
+                iframeRef={iframeRef}
+                outputRef={outputRef}
+              />
             </ResizablePanel>
           )}
         </ResizablePanelGroup>
