+++ src/components/playground/PlaygroundOutput.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+
+interface PlaygroundOutputProps {
+  language: string;
+  output: string;
+  iframeRef: React.RefObject<HTMLDivElement>;
+  outputRef: React.RefObject<HTMLDivElement>;
+}
+
+export const PlaygroundOutput: React.FC<PlaygroundOutputProps> = ({
+  language,
+  output,
+  iframeRef,
+  outputRef,
+}) => {
+  const isDark = document.documentElement.classList.contains('dark');
+
+  return (
+    <div className="h-full flex flex-col">
+      {language === 'html' ? (
+        <div 
+          ref={iframeRef} 
+          className={`w-full h-full ${isDark ? 'bg-[#1e1e1e]' : 'bg-white'}`} 
+        />
+      ) : (
+        <div 
+          ref={outputRef} 
+          className="w-full h-full p-4 font-mono text-sm overflow-auto bg-black text-white"
+        >
+          {output}
+        </div>
+      )}
+    </div>
+  );
+};
