+++ src/components/playground/constants.ts
@@ -1,0 +1,21 @@
+export const SUPPORTED_LANGUAGES = [
+  { value: 'javascript', label: 'JavaScript', extension: '.js' },
+  { value: 'typescript', label: 'TypeScript', extension: '.ts' },
+  { value: 'python', label: 'Python', extension: '.py' },
+  { value: 'java', label: 'Java', extension: '.java' },
+  { value: 'csharp', label: 'C#', extension: '.cs' },
+  { value: 'cpp', label: 'C++', extension: '.cpp' },
+  { value: 'go', label: 'Go', extension: '.go' },
+  { value: 'rust', label: 'Rust', extension: '.rs' },
+  { value: 'ruby', label: 'Ruby', extension: '.rb' },
+  { value: 'php', label: 'PHP', extension: '.php' },
+  { value: 'sql', label: 'SQL', extension: '.sql' },
+  { value: 'html', label: 'HTML', extension: '.html' },
+  { value: 'css', label: 'CSS', extension: '.css' },
+  { value: 'json', label: 'JSON', extension: '.json' },
+  { value: 'markdown', label: 'Markdown', extension: '.md' },
+  { value: 'hcl', label: 'Terraform', extension: '.tf' },
+  { value: 'bicep', label: 'Bicep', extension: '.bicep' },
+  { value: 'powershell', label: 'PowerShell', extension: '.ps1' },
+  { value: 'shell', label: 'Bash/Shell', extension: '.sh' },
+] as const;
